update test cases

+++ generic-scalacheck/shared/src/test/scala/com/peknight/generic/deriving/Shape.scala
@@ -1,5 +1,5 @@
 package com.peknight.generic.deriving
 
 enum Shape derives CanEqual:
-  case Rectangle(width: Int, height: Int)
-  case Circle(radius: Int)
+  case Rectangle(width: Double, height: Double)
+  case Circle(radius: Double)
